+++ R/freepst-package.R
@@ -1,11 +1,14 @@
 #' Read OST/PST Files
 #'
 #' An 'rJava' wrapper around 'java-libpst' <https://github.com/rjohnsondev/java-libpst>.
+#' The PST File format is used by Outlook for the storage of emails. Over the
+#' years many people have accumulated a large amount of important email and
+#' other information into these files. Tools are provided to enable folks to
+#' access and extract this information for archival or analysis.
 #'
 #' @md
 #' @name freepst
 #' @docType package
 #' @author Bob Rudis (bob@@rud.is)
 #' @import rJava
-#' @importFrom dplyr bind_rows
 NULL