+++ R/read-folder.R
@@ -1,39 +1,40 @@
-#' Get all messages in an OST/PST
+#' Get all messages in an OST/PST with nested folder metadata
 #'
 #' @param path path to OST/PST
 #' @return data frame
+#' @author Torben Tvedebrink
 #' @export
 #' @examples
-#' read_pst(system.file("extdata/example-2013.ost", package="freepst"))
-#' read_pst(system.file("extdata/dist-list.pst", package="freepst"))
-read_pst <- function(path) {
-
+#' read_folder(system.file("extdata/example-2013.ost", package="freepst"))
+#' read_folder(system.file("extdata/dist-list.pst", package="freepst"))
+read_folder <- function (path) {
   path <- path.expand(path)
-  if (!file.exists(path)) stop(sprintf("'%s' not found.", path), call.=FALSE)
+  if (!file.exists(path))
+    stop(sprintf("'%s' not found.", path), call. = FALSE)
   f <- new(J("com.pff.PSTFile"), path)
-
   rf <- f$getRootFolder()
-
   depth <- -1
-
   msgs <- list()
-
-  process_folder <- function(folder) {
-
+  process_folder <- function(folder, par_folder = "") {
     if (folder$hasSubfolders()) {
-      child_folders <- folder$getSubFolders()
-      lapply(as.list(child_folders), process_folder)
+      child_folders <- as.list(folder$getSubFolders())
+      folder_name <- folder$getDisplayName()
+      par_folder <- paste0(par_folder, "/", folder_name)
+      lapply(child_folders, process_folder, par_folder = par_folder)
     }
-
     if (folder$getContentCount() > 0) {
-
       repeat {
         email <- folder$getNextChild()
-        if (is.jnull(email)) break
-        if (email$getMessageClass() == "IPM.Microsoft.ScheduleData.FreeBusy") next
+        if (is.jnull(email))
+          break
+        if (email$getMessageClass() == "IPM.Microsoft.ScheduleData.FreeBusy"){
+          next
+        }
+        ## browser()
         tmp <- list(
+          parent_folder = par_folder %l0% "",
           folder = folder$getDisplayName() %l0% "",
-          sent_by = email$getSenderName() %l0% "",
+          sent_by = email$getSenderName()  %l0% "",
           sent_by_addr = email$getSenderEmailAddress() %l0% "",
           received_by = email$getReceivedByName() %l0% "",
           received_by_addr = email$getReceivedByAddress() %l0% "",
@@ -50,9 +51,7 @@ read_pst <- function(path) {
         )
         msgs <<- c(msgs, list(tmp))
       }
-
     }
-
   }
 
   process_folder(rf)
@@ -61,8 +60,13 @@ read_pst <- function(path) {
 
   class(msgs) <- c("tbl_df", "tbl", "data.frame")
 
-  msgs
+  msgs[["folder"]] <- paste(msgs[["parent_folder"]], msgs[["folder"]], sep = "/")
+  msgs[["parent_folder"]] <- NULL
 
-  # return(dplyr::bind_rows(msgs))
+  ## The top level echos "Top of Outlook data file" hence, this and leading slashes are removed before return.
+  msgs[["folder"]] <- sub("^.*Top of .*\\/", "", msgs[["folder"]])
+  msgs[["folder"]] <- sub("^//", "/", msgs[["folder"]])
+
+  msgs
 
 }