+++ NEWS.md
@@ -1,2 +1,7 @@
+0.2.0
+* `read_folder()` which provides nested folder metadata (helped by Torben Tvedebrink)
+* switch to {tinytest}
+* Apache license (to align with the java lib)
+
 0.1.0 
 * Initial release